+++ examples/registry_status/app.py
@@ -42,7 +42,7 @@ app.config.update(
 xxl_job = FlaskXXLJob(app)
 
 
-@xxl_job.on_run
+@xxl_job.on_run("registryStatusJobHandler")
 def handle_run(request):
     return XXLJobResponse.success()